requirements.txt
astroid==2.3.1
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cssselect==1.0.3
Django==2.2.2
feedfinder2==0.0.4
feedparser==5.2.1
google==2.0.2
gunicorn==20.0.4
idna==2.8
isort==4.3.21
jieba3k==0.35.1
lazy-object-proxy==1.4.2
lxml==4.3.4
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.4.3
Pillow==6.0.0
pylint==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-file==1.4.3
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tinysegmenter==0.3
tldextract==2.2.1
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==5.1.0
wrapt==1.11.2
